problemset-76279/boors-va-corona/requirements.txt
django-redis==4.12.1
djangorestframework==3.12.2
Django==3.*
gunicorn==20.0.4
django-decouple==2.1
